Favorites fix 2

+++ backend/api/views.py
@@ -59,12 +59,15 @@ class RecipeViewSet(viewsets.ModelViewSet):
             serializer.save(recipe=recipe, user=request.user)
             serializer = FavoriteAndShoppingCartSerializer(recipe)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-        favorite = Favorite.objects.get(
-            user=request.user,
-            recipe__id=pk
-        )
-        favorite.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.method == 'DELETE':
+            try:
+                favorite = Favorite.objects.get(
+                    user=request.user, recipe__id=pk
+                )
+                favorite.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            except Favorite.DoesNotExist:
+                return Response(status=status.HTTP_400_BAD_REQUEST)
 
     @action(methods=['POST', 'DELETE'], detail=True,
             permission_classes=[permissions.IsAuthenticated],
